pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vietlaw"
version = "0.1.0"
description = "Kho Du Lieu Phap Luat Viet Nam cho AI/RAG"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Trustydev212"},
]
keywords = ["vietnam", "law", "legal", "rag", "ai", "nlp", "vietnamese"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
]
dependencies = [
    "PyYAML>=6.0.1",
    "jsonschema>=4.22.0",
    "openai>=1.35.0",
    "chromadb>=0.5.0",
]

[project.optional-dependencies]
reranking = [
    "sentence-transformers>=3.0.0",
]
export = [
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]
all = [
    "vietlaw[reranking,export,dev]",
]

[project.urls]
Homepage = "https://github.com/trustydev212/vietlaw"
Repository = "https://github.com/trustydev212/vietlaw"
Issues = "https://github.com/trustydev212/vietlaw/issues"

[tool.setuptools.packages.find]
include = ["vietlaw*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
pythonpath = ["."]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.coverage.run]
source = ["vietlaw", "scripts"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 60
